+++ run.sh
@@ -42,21 +42,6 @@ if [ ! -z ${GF_AWS_PROFILES+x} ]; then
     chmod 600 ~grafana/.aws/credentials
 fi
 
-if [ ! -z "${GF_INSTALL_PLUGINS}" ]; then
-  OLDIFS=$IFS
-  IFS=','
-  for plugin in ${GF_INSTALL_PLUGINS}; do
-    IFS=$OLDIFS
-    if [[ $plugin =~ .*\;.* ]]; then
-        pluginUrl=$(echo "$plugin" | cut -d';' -f 1)
-        pluginInstallFolder=$(echo "$plugin" | cut -d';' -f 2)
-        gosu grafana grafana-cli --pluginUrl ${pluginUrl} --pluginsDir "${GF_PATHS_PLUGINS}" plugins install "${pluginInstallFolder}"
-    else
-        gosu grafana grafana-cli --pluginsDir "${GF_PATHS_PLUGINS}" plugins install ${plugin}
-    fi
-  done
-fi
-
 if [ "z$DONT_COPY_STOCK_DASHBOARDS"  = "z" ]; then
   echo "Deleting existing provisioning"
   rm -rf /etc/grafana/provisioning/*